# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SendCodeInput {
  phone: string;
}

export interface SendCodeResult {
  phoneCodeHash: string;
  sessionId: string;
}

export interface SignInInput {
  phone: string;
  code: string;
  phoneCodeHash: string;
  sessionId: string;
  /** @nullable */
  password?: string | null;
}

export interface TelegramAccount {
  phone: string;
  id: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  has2fa?: boolean;
}

export interface SuccessResult {
  success: boolean;
  message: string;
}

export interface Disable2faInput {
  password: string;
}

export interface LoginCodeResult {
  found: boolean;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  from?: string | null;
  /** @nullable */
  date?: string | null;
}

export interface TelegramSession {
  hash: string;
  deviceModel: string;
  platform: string;
  appName: string;
  dateCreated: string;
  dateActive: string;
  ip: string;
  country: string;
  current: boolean;
}

export interface TerminateSessionInput {
  hash: string;
}

export interface ChangeEmailInput {
  email: string;
}

export interface VerifyEmailInput {
  email: string;
  code: string;
}