# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Telegram Account Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: telegram
    description: Telegram account management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /telegram/send-code:
    post:
      operationId: sendCode
      tags: [telegram]
      summary: Send login code to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCodeInput"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendCodeResult"

  /telegram/sign-in:
    post:
      operationId: signIn
      tags: [telegram]
      summary: Sign in with code (and optional 2FA password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"

  /telegram/accounts:
    get:
      operationId: listAccounts
      tags: [telegram]
      summary: List all logged-in accounts
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramAccount"

  /telegram/accounts/{phone}:
    get:
      operationId: getAccount
      tags: [telegram]
      summary: Get account details
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
    delete:
      operationId: removeAccount
      tags: [telegram]
      summary: Remove/logout account from manager
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /telegram/accounts/{phone}/disable-2fa:
    post:
      operationId: disable2fa
      tags: [telegram]
      summary: Disable 2FA on account
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Disable2faInput"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /telegram/accounts/{phone}/login-code:
    get:
      operationId: getLoginCode
      tags: [telegram]
      summary: Get recent Telegram login code from messages
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Login code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginCodeResult"

  /telegram/accounts/{phone}/sessions:
    get:
      operationId: getSessions
      tags: [telegram]
      summary: Get active sessions for account
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramSession"

  /telegram/accounts/{phone}/terminate-session:
    post:
      operationId: terminateSession
      tags: [telegram]
      summary: Terminate a specific session
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminateSessionInput"
      responses:
        "200":
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /telegram/accounts/{phone}/terminate-all-sessions:
    post:
      operationId: terminateAllSessions
      tags: [telegram]
      summary: Terminate all other sessions
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All sessions terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /telegram/accounts/{phone}/change-email:
    post:
      operationId: changeEmail
      tags: [telegram]
      summary: Request email change (sends verification code)
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEmailInput"
      responses:
        "200":
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /telegram/accounts/{phone}/verify-email:
    post:
      operationId: verifyEmail
      tags: [telegram]
      summary: Verify email change with code
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SendCodeInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string

    SendCodeResult:
      type: object
      required: [phoneCodeHash, sessionId]
      properties:
        phoneCodeHash:
          type: string
        sessionId:
          type: string

    SignInInput:
      type: object
      required: [phone, code, phoneCodeHash, sessionId]
      properties:
        phone:
          type: string
        code:
          type: string
        phoneCodeHash:
          type: string
        sessionId:
          type: string
        password:
          type: ["string", "null"]

    TelegramAccount:
      type: object
      required: [phone, id, firstName]
      properties:
        phone:
          type: string
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        has2fa:
          type: boolean

    SuccessResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    Disable2faInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    LoginCodeResult:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        code:
          type: ["string", "null"]
        from:
          type: ["string", "null"]
        date:
          type: ["string", "null"]

    TelegramSession:
      type: object
      required: [hash, deviceModel, platform, appName, dateCreated, dateActive, ip, country, current]
      properties:
        hash:
          type: string
        deviceModel:
          type: string
        platform:
          type: string
        appName:
          type: string
        dateCreated:
          type: string
        dateActive:
          type: string
        ip:
          type: string
        country:
          type: string
        current:
          type: boolean

    TerminateSessionInput:
      type: object
      required: [hash]
      properties:
        hash:
          type: string

    ChangeEmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string

    VerifyEmailInput:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
        code:
          type: string