# lib/api-zod/src/generated/types/loginCodeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginCodeResult {
  found: boolean;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  from?: string | null;
  /** @nullable */
  date?: string | null;
}