# lib/api-zod/src/generated/types/sendCodeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SendCodeResult {
  phoneCodeHash: string;
  sessionId: string;
}