# lib/api-zod/src/generated/types/signInInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SignInInput {
  phone: string;
  code: string;
  phoneCodeHash: string;
  sessionId: string;
  /** @nullable */
  password?: string | null;
}