# lib/api-zod/src/generated/types/telegramAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramAccount {
  phone: string;
  id: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  has2fa?: boolean;
}