# lib/api-zod/src/generated/types/telegramSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramSession {
  hash: string;
  deviceModel: string;
  platform: string;
  appName: string;
  dateCreated: string;
  dateActive: string;
  ip: string;
  country: string;
  current: boolean;
}