# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram Account Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyEmailInput {
  email: string;
  code: string;
}